+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/extendablebutton/StartNavigationButtonComponentTest.kt
@@ -1,14 +1,10 @@
 package com.mapbox.navigation.dropin.extendablebutton
 
-import android.view.ViewGroup
-import android.widget.LinearLayout
+import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import com.mapbox.geojson.Point
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
-import com.mapbox.navigation.dropin.MapboxExtendableButtonParams
-import com.mapbox.navigation.dropin.R
-import com.mapbox.navigation.dropin.internal.extensions.recreateButton
 import com.mapbox.navigation.dropin.util.TestStore
 import com.mapbox.navigation.dropin.util.TestingUtil
 import com.mapbox.navigation.testing.MainCoroutineRule
@@ -16,20 +12,16 @@ import com.mapbox.navigation.ui.app.internal.State
 import com.mapbox.navigation.ui.app.internal.destination.Destination
 import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.app.internal.navigation.NavigationStateAction
+import com.mapbox.navigation.ui.app.internal.routefetch.RouteOptionsProvider
 import com.mapbox.navigation.ui.app.internal.routefetch.RoutePreviewAction
 import com.mapbox.navigation.ui.app.internal.routefetch.RoutePreviewState
 import com.mapbox.navigation.ui.base.view.MapboxExtendableButton
 import io.mockk.every
 import io.mockk.mockk
-import io.mockk.mockkStatic
 import io.mockk.spyk
-import io.mockk.unmockkAll
 import io.mockk.verify
-import io.mockk.verifyOrder
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runBlockingTest
-import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -43,39 +35,21 @@ class StartNavigationButtonComponentTest {
     @get:Rule
     val coroutineRule = MainCoroutineRule()
 
-    private val layoutParams: LinearLayout.LayoutParams = mockk(relaxed = true)
-    private val initialParams =
-        MapboxExtendableButtonParams(R.style.DropInStyleExitButton, layoutParams)
-    private val updatedParams =
-        MapboxExtendableButtonParams(R.style.DropInStyleRecenterButton, layoutParams)
-
-    private val store = spyk(TestStore())
-    private val button = spyk(MapboxExtendableButton(ApplicationProvider.getApplicationContext()))
-    private val buttonParams = MutableStateFlow(initialParams)
-    private val buttonContainer = mockk<ViewGroup>(relaxed = true)
-    private val mapboxNavigation = mockk<MapboxNavigation>(relaxed = true)
-    private val sut = StartNavigationButtonComponent(store, buttonContainer, buttonParams, mockk())
+    private lateinit var store: TestStore
+    private lateinit var mapboxNavigation: MapboxNavigation
+    private lateinit var routeOptionsProvider: RouteOptionsProvider
+    private lateinit var button: MapboxExtendableButton
+    private lateinit var sut: StartNavigationButtonComponent
 
     @Before
     fun setUp() {
-        mockkStatic("com.mapbox.navigation.dropin.internal.extensions.ViewGroupEx")
-        every { buttonContainer.recreateButton(any()) } returns button
-    }
+        val context = ApplicationProvider.getApplicationContext<Context>()
+        store = spyk(TestStore())
+        mapboxNavigation = mockk(relaxed = true)
+        routeOptionsProvider = mockk()
+        button = spyk(MapboxExtendableButton(context))
 
-    @After
-    fun cleanUp() {
-        unmockkAll()
-    }
-
-    @Test
-    fun `should observe params and recreate button`() = runBlockingTest {
-        sut.onAttached(mapboxNavigation)
-        buttonParams.value = updatedParams
-
-        verifyOrder {
-            buttonContainer.recreateButton(initialParams)
-            buttonContainer.recreateButton(updatedParams)
-        }
+        sut = StartNavigationButtonComponent(store, routeOptionsProvider, button)
     }
 
     @Test